warehouse: Separate config from node.Config to warehouse.Config
